Movement and camera attempt #1
for movement the x axis works but i can figure out why the y axis of movement has a "lock" on it. it never goes higher than 1 or -1.

the camera doesnt show up when playing the game but in the mini view it shows it working? i had it working for a second then it didnt work again.

im using visual scripting to figure out the movement. since i dont know real boy coding.